creating seeds for food categories and food items

+++ database/migrations/2020_02_29_194418_create_food_items_table.php
@@ -18,7 +18,7 @@ class CreateFoodItemsTable extends Migration
             $table->string('title');
             $table->text('description');
             $table->text('image_url');
-            $table->integer('price');
+            $table->float('price',8, 2);
             $table->bigInteger('category_id')->unsigned()->nullable();
             $table->timestamps();
         });
